refactor: decouple tightly coupled components and hooks
